use statsd exporter as an example
